+++ assets/plugins/wwsu-calendar/js/wwsu-calendar-web.js
@@ -802,10 +802,7 @@ class CalendarDb {
 			}
 
 			// Duplicate schedule for the same event? That's a problem! Throw an error.
-			if (
-				!overridden.scheduleType &&
-				overridden.calendarID === overrides.calendarID
-			) {
+			if (overridden.calendarID === overrides.calendarID) {
 				throw new Error(
 					`There is a schedule for calendar ID ${overridden.calendarID} that would overlap another schedule for the same calendar ID. This is not allowed.`
 				);